Major Changes to GUI
The old GUI has been deleted in favor of a new GUI I have made. There is
now a graph page and sensor manager page.

The structure of the gui's frames are now in their own classes which
makes the program more flexable. Within the sensor_manager there is now
a dynamic list that forms when a sensor is connected.