New Sensor Features / Upcoming Graph Features
This is a commit in the middle of changes.

There is now voltage control on the GUI as well as faster plotting. Each
channel is now plotted with the same line and dot lines but using only
one plot command making the GUI faster. Also the voltage can now be
incremented +5 or -5 (integer value). There is also one placeholder
sensor when GUI starts as well as placeholder text that will be replaced
with graph controls within the next few commits.